requirements.txt
pandas
numpy
matplotlib
scipy
scikit-learn
scikit-image
Pillow
requests
pydantic